write resolvers for new fields + queries
