+++ server/src/resolvers/Player.js
@@ -1,4 +1,5 @@
-const { sportsFeedRequest } = require('../sportsFeed')
+const { sportsFeedRequest } = require('../sportsFeed/api')
+const { getStartDate, getEndDate, parseDate } = require('../sportsFeed/helpers')
 
 function fetchPlayerStats(playerId) {
 	return sportsFeedRequest(`2018-2019-regular/player_stats_totals.json?player=${playerId}`)
@@ -54,10 +55,27 @@ function fetchPlayerStats(playerId) {
 		})
 }
 
+function calculateGameCount(teamId, startDate, endDate) {
+	return sportsFeedRequest(
+		`2018-2019-regular/team_gamelogs.json?team=${teamId}&date=from-${startDate}-to-${endDate}`
+	)
+		.then(res => res.json())
+		.then(json => json.gamelogs.length)
+}
+
 function stats(parent, args) {
 	return fetchPlayerStats(parent.id)
 }
 
+function gameCountThisWeek(parent, args) {
+	return calculateGameCount(
+		parent.currentTeam.id,
+		parseDate(getStartDate()),
+		parseDate(getEndDate(getStartDate()))
+	)
+}
+
 module.exports = {
 	stats,
+	gameCountThisWeek,
 }
